+++ app/frontend/components/images-view/index.ts
@@ -1,0 +1,1 @@
+export { ImagesView } from './images-view';
