+++ app/frontend/components/types.ts
@@ -1,0 +1,4 @@
+import { RouterOutput } from '../utils/trpc';
+
+export type Images = RouterOutput['fetchCurrentPage']['images'];
+export type Image = Images[number];
